Refactor code to be cleaner.
Bench: 2886048
